restrict question update and delete by unauthorized user using gate

+++ app/Http/Controllers/QuestionController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\QuestionStoreRequest;
 use App\Question;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionController extends Controller
 {
@@ -72,6 +73,9 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
+        if (Gate::denies('update-question', $question)) {
+            abort(403, "Access Denied");
+        }
         return view('questions.edit', compact('question'));
     }
 
@@ -84,6 +88,9 @@ class QuestionController extends Controller
      */
     public function update(QuestionStoreRequest $request, Question $question)
     {
+        if (Gate::denies('update-question', $question)) {
+            abort(403, "Access Denied");
+        }
         $question->update([
             'title' => $request->title,
             'body' => $request->body,
@@ -99,6 +106,9 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (Gate::denies('delete-question', $question)) {
+            abort(403, "Access Denied");
+        }
         $question->delete();
         return redirect(route('questions.index'))->with('success', "Your question has been deleted");
     }
